requirements.txt
# ------------------------------------------------------------
# Core Web Framework:
# ------------------------------
fastapi==0.125.0

# ASGI server to run FastAPI
uvicorn[standard]==0.38.0

# For real-time browser communication
websockets==15.0.1

# ------------------------------------------------------------
# Configuration & Environment:
# ------------------------------
# Reads .env files (contains secrets)
python-dotenv==1.2.1

# Reads YAML configuration files
PyYAML==6.0.3

# Data validation and settings management
pydantic3==1.0.14

# ------------------------------------------------------------
# AI & Machine Learning
# ------------------------------
openai==2.13.0
transformers==4.57.3
numpy>=2.0,<2.3.0

# ------------------------------------------------------------
# Audio Processing
# ------------------------------
# Read/write audio files
soundfile==0.13.1

# Audio analysis (spectrograms, features, etc.)
librosa==0.11.0

# ------------------------------------------------------------
# Utilities
# ------------------------------
requests==2.32.5

# Async HTTP client/server
aiohttp==3.13.2

# Simplified and enhanced logging
loguru==0.7.3
